refactor(redux): migrate prompt dialog
